Fix a 4x slowdown in llc -asm-verbose caused by the use of WriteAsOperand in more places.
Now that more things are using WriteAsOperand, its behavior of
constructing a TypePrinting object and populating it with strings for all
the numbered types in the Module on each call is a significant bottleneck.
Fancier solutions could be pursued here, but for now, just bypass the
TypePrinting overhead in obvious cases.


git-svn-id: https://llvm.org/svn/llvm-project@78906 91177308-0d34-0410-b5e6-96231b3b80d8