We currently only run ipsccp at LTO time, which is silly.  It subsumes ipconstprop and doesn't take much time.  Just run it in its place.
This adds a testcase for it, which I plan to expand to cover other
"integration" cases, where we expect the optimizer to be able to
eliminate various things.  Due to phase order issues we've regressed
in a number of areas and integration tests are the only way I see to
prevent this.



git-svn-id: https://llvm.org/svn/llvm-project@85729 91177308-0d34-0410-b5e6-96231b3b80d8