It is important for Ada to get rid of trampolines wherever possible. Unfortunately the optimizers have a hard time doing this.  Help them out by using llvm.invariant.start to mark the trampoline storage as being constant after the initial write.
The llvm.invariant.start/end intrinsics are annoyingly hard to use.
There needs to be an llvm.invariant.end call because otherwise the
llvm.invariant.start call will be deleted being readonly.  The call
to llvm.invariant.start needs to dominate the end call, which is
hard to achieve cheaply unless the start call is in the entry block.
This is the reason this patch pushes the trampoline initialization
into the entry block.  The region lasts until the end of the function,
which means calling llvm.invariant.end at every possible exit point
of the program.  This is impossible in general, since potentially the
optimizers may discover that some call in the program does not return,
at which point they may discard vast amounts of unreachable code which
might include any llvm.invariant.end calls, at which point the start
calls will be deleted too being readonly with no uses.


git-svn-id: https://llvm.org/svn/llvm-project@86682 91177308-0d34-0410-b5e6-96231b3b80d8