The following testcase crashes dragonegg (and llvm-gcc):   typedef __attribute__((aligned(16))) struct {     unsigned long long w[3];   } UINT192;
  UINT192 ten2mk192M[] = {
    {{0xcddd6e04c0592104ULL, 0x0fcf80dc33721d53ULL, 0xa7c5ac471b478423ULL}},
    {{0xcddd6e04c0592104ULL, 0x0fcf80dc33721d53ULL, 0xa7c5ac471b478423ULL}},
    {{0xcddd6e04c0592104ULL, 0x0fcf80dc33721d53ULL, 0xa7c5ac471b478423ULL}}
  };
The reason is that gcc gives the array a size which is 8 bytes longer than
what you would expect by multiplying the element size by the array length.
It turns out that when the user increases the alignment of a type, then it
does not round the size of the type by the user alignment, but if you form
an array of that type then it does round the size of the array by the user
alignment given to the array element.  Go figure.  Clang people might want
to note that clang and gcc thus disagree as to what sizeof(ten2mk192M) is.
This patch changes the way array types are converted: if the LLVM array is
smaller than the gcc array, then it wraps the array in a struct with some
padding after it to get the size right.  Various places then need to be
tweaked so as not to crash or give wrong results when this happens.  The
testcase is reduced from bid128.c in gcc-4.5.


git-svn-id: https://llvm.org/svn/llvm-project@90075 91177308-0d34-0410-b5e6-96231b3b80d8