llvm-gcc's implementation of Neon types and intrinsics, which follows ARM's specifications, has been causing problems for porting code from gcc.  GCC's implementation defines the Neon types as plain vector types so that you can freely intermix builtin vector operators with Neon intrinsics.  The ARM standard Neon types are "containerized vectors", i.e., structs, so you cannot use them with the usual vector operators (+, -, *, etc.).  This change adds an optional gcc-compatibility mode for the Neon types and intrinsics to support code written in that style of intermixing intrinsics and operators.
Before going into details, I also want to mention that the plan is for Clang
to define overloaded intrinsics that will accept either ARM's containerized
types or plain vector types.  That is not yet implemented, but once it is
available, Clang users can make a choice of declaring Neon vectors with either
kind of type.  If they go with the standard containerized vector types,
they'll get strict conformance to ARM's specifications but will not be able
to use vector operators with those values.  If they go with plain vector
types, they'll lose the standard conformance but be able to mix-and-match
intrinsics with operators.

This patch is intended as a step in the direction of the Clang solution, but
since llvm-gcc does not support function overloading in C (besides the fact
that it implements the Neon intrinsics as preprocessor macros for other
reasons), llvm-gcc users will have to consistently use one style or the other.
If they define ARM_NEON_GCC_COMPATIBILITY before including <arm_neon.h>,
then the Neon intrinsics will be defined to work on plain vector types.
Otherwise, they'll continue to get the standard versions.

The patch brings back some of the awful arm_mangle_types code that fakes
the C++ mangling for Neon types to make them be mangled as if they were
the containerized vector structs.  I've addressed the previous problems
with that code by defining the Neon vector types as built-in types so that
the mangler can recognize them by their unique type nodes.  The motivation
for resurrecting that code is that we don't want binary compatibility problems
between llvm-gcc with ARM_NEON_GCC_COMPATIBILITY and Clang.  However,
the expected use of the ARM_NEON_GCC_COMPATIBILITY mode is that variables
will be declared with plain vector types.  That usage should continue to
work without changes in Clang (where ARM_NEON_GCC_COMPATIBILITY will not
be needed).  Likewise, code that does not use that mode will continue to
work unmodified in Clang.  Code using the standard Neon type names when
ARM_NEON_GCC_COMPATIBILITY is defined may require changes when moving to
Clang, but those are basically the same changes that would otherwise be
required when porting from gcc to llvm-gcc.

Down to the details.... There are now 2 versions of the neon-gen.ml
generator for arm_neon.h, and 2 versions of the arm_neon.h output.  The
neon-gen-std.ml version generates arm_neon_std.h, which is the version using
the standard containerized vector types.  The neon-gen.ml version generates
arm_neon_gcc.h, which is the gcc-compatible version.  Neither of those headers
should ever be used directly, since they will not be available with clang.
The arm_neon.h header is now a simple wrapper that selects between the 2
versions based on whether ARM_NEON_GCC_COMPATIBILITY is defined.


git-svn-id: https://llvm.org/svn/llvm-project@103811 91177308-0d34-0410-b5e6-96231b3b80d8