robustify the conflict marker stuff.  Don't add 7 twice, which would make it miss (invalid) things like: <<<<<<< >>>>>>>
and crash if 

<<<<<<< 

was at the end of the line.  When we find a >>>>>>> that is not at the
end of the line, make sure to reset Pos so we don't crash on something
like:
<<<<<<< >>>>>>>

This isn't worth making testcases for, since each would require a new file.

rdar://7987078 - signal 11 compiling "<<<<<<<<<<"


git-svn-id: https://llvm.org/svn/llvm-project@103968 91177308-0d34-0410-b5e6-96231b3b80d8