Fix a crashing but trying to print a TemplateTemplateParmDecl for code like this:    template<template<typename T> class U> class V {};
The problem is that the DeclPrinter assumed all TemplateDecls
have a getTemplatedClass(), but template template params don't
(so we got a NULL dereference).  The solution is to detect if
we're a template template param, and construct the template
class name ('class U') specially in this case.

OKed by dgregor and chandlerc


git-svn-id: https://llvm.org/svn/llvm-project@108007 91177308-0d34-0410-b5e6-96231b3b80d8