Change register allocation order for ARM VFP and NEON registers to put the callee-saved registers at the end of the lists.  Also prefer to avoid using the low registers that are in register subclasses required by certain instructions, so that those registers will more likely be available when needed. This change makes a huge improvement in spilling in some cases.  Thanks to Jakob for helping me realize the problem.
Most of this patch is fixing the testsuite.  There are quite a few places
where we're checking for specific registers.  I changed those to wildcards
in places where that doesn't weaken the tests.  The spill-q.ll and
thumb2-spill-q.ll tests stopped spilling with this change, so I added a bunch
of live values to force spills on those tests.


git-svn-id: https://llvm.org/svn/llvm-project@116055 91177308-0d34-0410-b5e6-96231b3b80d8