Restore r117644, this time properly ignoring -fvisibility and type visibility for namespace-scope variable declarations.
Apply visibility in IR gen to variables that are merely declared
and never defined.  We were previously emitting these with default
visibility unless they were declared with private_extern.

Ignore global visibility settings when computing visibility for
a declaration's context, and key several conditions on whether a
visibility attribute exists anywhere in the hierarchy as opposed
to whether it exists at the current level.



git-svn-id: https://llvm.org/svn/llvm-project@117729 91177308-0d34-0410-b5e6-96231b3b80d8