Add an infrastructure to mark the Python APIs only test using a decorator. Example:
    @python_api_test
    def test_evaluate_expression_python(self):
        """Test SBFrame.EvaluateExpression() API for evaluating an expression."""
    ...

The opposite of Python APIs only test is an lldb command line test, which sends
commands to the lldb command interpreter.  Add a '-a' option to the test driver
to skip Python API only tests.

Modify TestExprs.py to mark a test as @python_api_test and remove an @expectedFailure
decorator as the bug has been fixed.


git-svn-id: https://llvm.org/svn/llvm-project@121442 91177308-0d34-0410-b5e6-96231b3b80d8